docs: generate docs, add more ts examples, remove js examples
