--- examples/javascript/.prettierrc
@@ -1,6 +1,0 @@
-{
-  "jsxBracketSameLine": true,
-  "printWidth": 70,
-  "singleQuote": true,
-  "trailingComma": "es5"
-}
