--- examples/javascript/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "javascript-papercut-examples",
-  "version": "0.0.0",
-  "type": "module",
-  "author": "Armando Magalhães",
-  "license": "MIT",
-  "engines": {
-    "node": "^16"
-  },
-  "scripts": {
-    "run-scraper": "node --experimental-json-modules --max-old-space-size=8192",
-    "lint": "prettier --check './**/*.js'",
-    "lint:fix": "prettier --write './**/*.js'"
-  },
-  "dependencies": {
-    "@armand1m/papercut": "file:../.."
-  },
-  "devDependencies": {
-    "prettier": "^2.3.2"
-  }
-}
